+++ halatuju_api/halatuju/settings/base.py
@@ -91,6 +91,9 @@ REST_FRAMEWORK = {
     'DEFAULT_PARSER_CLASSES': [
         'rest_framework.parsers.JSONParser',
     ],
+    'DEFAULT_AUTHENTICATION_CLASSES': [
+        'halatuju.middleware.supabase_auth.SupabaseAuthentication',
+    ],
     'DEFAULT_PERMISSION_CLASSES': [
         'halatuju.middleware.supabase_auth.SupabaseIsAuthenticated',
     ],